repo, usecase, service layer for chat & proto file

+++ internal/entity/message.go
@@ -3,9 +3,9 @@ package entity
 import "time"
 
 type Message struct {
-	ID        int
-	ChatID    int
-	Sender    int
+	ID        int64
+	ChatID    int64
+	Sender    int64
 	Content   string
 	CreatedAt time.Time
 	UpdateAt  time.Time
